+++ src/num/unit/bi.rs
@@ -365,7 +365,7 @@ impl UnitBi {
         // Ensure the threshold is positive and above a small epsilon value
         // in order to prevent infinite loops
         let effective_threshold =
-            if threshold <= 0.0 { crate::ExtFloatConst::MEDIUM_MARGIN } else { threshold };
+            if threshold <= 0.0 { crate::FloatConst::MEDIUM_MARGIN } else { threshold };
 
         while remainder.abs() > effective_threshold {
             let (size, unit) = Self::reduce(remainder);
@@ -499,7 +499,7 @@ mod tests {
     #[cfg(any(feature = "std", all(feature = "alloc", feature = "_float_f64")))]
     use crate::vec_ as vec;
     #[allow(unused_imports)]
-    use crate::ExtFloatConst;
+    use crate::FloatConst;
 
     /* reduce */